use int (for now...)

+++ src/users/application/UserService.go
@@ -12,7 +12,7 @@ func NewUserService(userRepository domain.UserRepository) *UserService {
 	return &UserService{userRepository: userRepository}
 }
 
-func (service *UserService) GetUser(id int64) UserDto {
+func (service *UserService) GetUser(id int) UserDto {
 	user := service.userRepository.GetUser(id)
 	userDto := UserDto{Id: user.Id, Name: user.Name, Email: user.Email}
 	return userDto
